pyproject.toml
[tool.poetry]
name = "ragquery"
version = "0.1.0"
description = ""
authors = ["suraj chapagain <surajchapagain7743@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.39.0"
python-dotenv = "^1.0.1"
llama-index = "^0.11.19"
llama-index-readers-web = "^0.2.4"
simpledocumentstore = "^0.1"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
requests = "^2.32.3"
pydantic = "^2.9.2"
termcolor = "^2.5.0"
firecrawl-py = "0.0.20"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
